@@ -65,7 +65,7 @@ extras_require['all'] = [y for x in extras_require.values() for y in x]
 setup(
     name='replicat',
     version='0.0.1',
-    python_requires=">=3.8",
+    python_requires=">=3.9",
     description='Configurable and lightweight backup utility with '
     'deduplication, encryption and stuff.',
     packages=find_packages(),
@@ -73,7 +73,6 @@ setup(
         'httpx',
         'cryptography',
         'backoff',
-        'pybind11',
         'appdirs',
         'tqdm',
         'sty',
