@@ -1,5 +1,7 @@
 import os
+from collections import Counter
 from hashlib import blake2b, scrypt
+from random import Random
 
 import pytest
 from cryptography.hazmat.primitives.ciphers.aead import AESGCM, ChaCha20Poly1305
@@ -9,6 +11,13 @@ from replicat.utils import adapters
 
 
 class TestAESGCM:
+    @pytest.mark.parametrize('key_bits', [128, 192, 256])
+    def test_generate_key(self, key_bits):
+        adapter = adapters.aes_gcm(key_bits=key_bits, nonce_bits=96)
+        key = adapter.generate_key()
+        assert isinstance(key, bytes)
+        assert len(key) * 8 == key_bits
+
     def test_encrypt(self):
         adapter = adapters.aes_gcm(key_bits=256, nonce_bits=96)
         key = b'<key>'.ljust(32, b'\x00')
@@ -43,6 +52,12 @@ class TestAESGCM:
 
 
 class TestChaCha20Poly1305:
+    def test_generate_key(self):
+        adapter = adapters.chacha20_poly1305()
+        key = adapter.generate_key()
+        assert isinstance(key, bytes)
+        assert len(key) * 8 == 256
+
     def test_encrypt(self):
         adapter = adapters.chacha20_poly1305()
         key = b'<key>'.ljust(32, b'\x00')
@@ -79,23 +94,74 @@ class TestChaCha20Poly1305:
 
 
 class TestScrypt:
-    def test_derive(self):
+    def test_generate_derivation_params(self):
         adapter = adapters.scrypt(length=17, n=4, r=8, p=1)
-        params = adapter.derivation_params()
+        params = adapter.generate_derivation_params()
+        assert isinstance(params, bytes)
+        assert len(params) == 17
+
+    def test_derive_without_context(self):
+        adapter = adapters.scrypt(length=17, n=4, r=8, p=1)
+        params = adapter.generate_derivation_params()
         key = adapter.derive(b'<password>', params=params)
         assert key == scrypt(b'<password>', n=4, r=8, p=1, dklen=17, salt=params)
 
+    def test_derive_with_context(self):
+        adapter = adapters.scrypt(length=17, n=4, r=8, p=1)
+        params = adapter.generate_derivation_params()
+        key = adapter.derive(
+            b'<password>', context=b'<additional context>', params=params
+        )
+        assert key == scrypt(
+            b'<password>',
+            n=4,
+            r=8,
+            p=1,
+            dklen=17,
+            salt=params + b'<additional context>',
+        )
+
 
 class TestBlake2B:
-    def test_derive(self):
+    def test_generate_derivation_params(self):
         adapter = adapters.blake2b(length=17)
-        params = adapter.derivation_params()
-        key = adapter.derive(b'<password>', params=params)
-        assert key == blake2b(b'<password>', digest_size=17, key=params).digest()
+        params = adapter.generate_derivation_params()
+        assert isinstance(params, bytes)
+        assert len(params) == blake2b.SALT_SIZE
+
+    def test_derive_without_context(self):
+        adapter = adapters.blake2b(length=17)
+        params = adapter.generate_derivation_params()
+        key = adapter.derive(b'<key material>', params=params)
+        assert (
+            key == blake2b(key=b'<key material>', salt=params, digest_size=17).digest()
+        )
+
+    def test_derive_with_context(self):
+        adapter = adapters.blake2b(length=17)
+        params = adapter.generate_derivation_params()
+        key = adapter.derive(
+            b'<key material>', context=b'<additional context>', params=params
+        )
+        assert (
+            key
+            == blake2b(
+                b'<additional context>',
+                key=b'<key material>',
+                salt=params,
+                digest_size=17,
+            ).digest()
+        )
+
+    def test_generate_mac_params(self):
+        adapter = adapters.blake2b(length=17)
+        params = adapter.generate_mac_params()
+        assert isinstance(params, bytes)
+        assert len(params) == blake2b.MAX_KEY_SIZE
 
     def test_mac(self):
         adapter = adapters.blake2b(length=17)
-        params = adapter.mac_params()
+        params = adapter.generate_mac_params()
         mac = adapter.mac(b'<data>', params=params)
         assert mac == blake2b(b'<data>', digest_size=17, key=params).digest()
 
@@ -117,7 +183,11 @@ class TestGCLMULChunker:
             (5, 10, [b'\xaa' * 12], [6, 6]),
             (5, 10, [b'\xaa' * 13], [6, 7]),
             (5, 10, [b'\xaa' * 14], [7, 7]),
-            (5, 10, [b'\xaa' * 15], [7, 8]),
+            (5, 10, [b'\xaa' * 15], [10, 5]),
+            (5, 10, [b'\xaa' * 16], [10, 6]),
+            (5, 10, [b'\xaa' * 17], [10, 7]),
+            (5, 10, [b'\xaa' * 18], [10, 8]),
+            (5, 10, [b'\xaa' * 19], [10, 9]),
             (4, 4, [b'\xaa'] * 20, [4] * 5),
             (4, 4, [b'\xaa' * 20], [4] * 5),
             (10, 12, [b'\xaa'] * 11, [11]),
@@ -130,19 +200,25 @@ class TestGCLMULChunker:
         assert [len(x) for x in chunks] == list(sizes)
 
     def test_personalization(self):
-        data = os.urandom(1_000_000)
+        rnd = Random(0)
+        data = rnd.randbytes(1_000_000)
+        person = bytearray(rnd.randbytes(16))
         chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
-        chunks = list(chunker([data]))
+        chunks = list(chunker([data], params=bytes(person)))
 
+        person[0] = (person[0] - 1) % 255
         person_chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
-        person_chunks = list(person_chunker([data], params=os.urandom(16)))
+        person_chunks = list(person_chunker([data], params=bytes(person)))
 
         assert person_chunks != chunks
         assert b''.join(chunks) == b''.join(person_chunks) == data
 
-    @pytest.mark.parametrize('person', [None, b'', b'abcd', os.urandom(64)])
-    def test_sequence_stabilizes(self, person):
-        data = bytearray(os.urandom(1_000_000))
+    # Hand-picked values
+    @pytest.mark.parametrize('seed', [507, 11219, 25750, 31286])
+    def test_sequence_stabilizes(self, seed):
+        rnd = Random(seed)
+        data = bytearray(rnd.randbytes(1_000_000))
+        person = rnd.randbytes(16)
         before_chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
         before_chunks = list(before_chunker([data], params=person))
 
@@ -150,19 +226,46 @@ class TestGCLMULChunker:
         after_chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
         after_chunks = list(after_chunker([data], params=person))
 
-        assert before_chunks[-1] == after_chunks[-1]
+        i, j = len(before_chunks), len(after_chunks)
+        while before_chunks[i - 1] == after_chunks[j - 1]:
+            i -= 1
+            j -= 1
 
-    @pytest.mark.parametrize('person', [None, b'', b'abcd', os.urandom(64)])
-    @pytest.mark.parametrize('repeating_bytes', [1_500, 5_000, 20_000, 50_000])
-    def test_repetition(self, person, repeating_bytes):
-        data = os.urandom(repeating_bytes)
-        counts = []
+        assert 0 < i < 4
+        assert 0 < j < 4
+        assert b''.join(before_chunks[:i])[1:] == b''.join(after_chunks[:j])[1:]
+        assert data.startswith(b''.join(after_chunks[:j]))
 
-        for i in (10, 20, 30):
-            chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
-            chunks = list(chunker([data] * i))
-            counts.append((len(chunks), len(set(chunks))))
+    @pytest.mark.parametrize(
+        'seed, size', [(0, 1_001), (1, 2_000), (2, 497), (2, 4_023), (3, 5_001)]
+    )
+    def test_repetition(self, seed, size):
+        rnd = Random(seed)
+        repeated_data = bytearray(rnd.randbytes(size))
+        person = rnd.randbytes(16)
 
-        first, second, third = counts
-        assert first[0] < second[0] < third[0]
-        assert first[1] == second[1] == second[1]
+        chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
+        chunks_50 = list(chunker([repeated_data] * 50, params=person))
+        assert b''.join(chunks_50) == repeated_data * 50
+
+        # Establish a pattern
+        counts_50 = Counter()
+        for x in chunks_50:
+            counts_50[x] += 1
+
+        chunk_pattern = [k for k, v in counts_50.items() if v > 1]
+        assert 0 < len(chunk_pattern) < 4
+        assert 1 < len(counts_50) - len(chunk_pattern) < 4
+
+        # Make sure we observe the same pattern in longer sequences
+        chunks_100 = list(chunker([repeated_data] * 100, params=person))
+        assert b''.join(chunks_100) == repeated_data * 100
+        unique_chunks_100 = set(chunks_100)
+        assert unique_chunks_100.issuperset(chunk_pattern)
+        assert len(unique_chunks_100) - len(chunk_pattern) < 4
+
+    def test_generate_chunking_params(self):
+        chunker = adapters.gclmulchunker(min_length=500, max_length=10_000)
+        params = chunker.generate_chunking_params()
+        assert isinstance(params, bytes)
+        assert len(params) == 16
